Create initial ols queries
