+++ traitcuration/traits/datasources/zooma.py
@@ -8,8 +8,9 @@ import logging
 from django.db import transaction
 
 from ..models import Trait, MappingSuggestion, OntologyTerm, User, Status, Mapping
-from .ols import make_ols_query, get_ontology_id
+from .ols import make_ols_query, get_ontology_id, get_term_status
 from .oxo import make_oxo_query
+
 
 logging.basicConfig()
 logger = logging.getLogger('ZOOMA')
@@ -106,21 +107,6 @@ def create_local_term(suggested_term_iri):
     term = OntologyTerm(curie=term_info['curie'], iri=suggested_term_iri, label=term_info['label'], status=term_status)
     term.save()
     return term
-
-
-def get_term_status(is_obsolete, ontology_id=None):
-    """
-    Takes the ontology_id of a term and a whether it is obsolete or not, and returns its calculated status.
-    'Obsolete' if the is_obsolete flag is true, 'Deleted' if no info was found in any ontology, 'Current' if its
-    ontology is EFO, and 'Needs Import' if info about the term was found in another ontology.
-    """
-    if ontology_id is None:
-        return Status.DELETED
-    if is_obsolete:
-        return Status.OBSOLETE
-    if ontology_id == 'efo':
-        return Status.CURRENT
-    return Status.NEEDS_IMPORT
 
 
 @transaction.atomic
